requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
slowapi
requests
celery
# openai
groq
redis